[Patch] Updated the create project cli
