+++ scripts/install-untoldengine-create.sh
@@ -98,14 +98,14 @@ if command -v $CLI_NAME &> /dev/null; then
     echo -e "  ${CYAN}$CLI_NAME --help${NC}"
     echo ""
     echo "Examples:"
-    echo -e "  ${CYAN}# Create macOS project${NC}"
-    echo -e "  ${CYAN}$CLI_NAME create MyGame${NC}"
+    echo -e "  ${CYAN}# Create a macOS project${NC}"
+    echo -e "  ${CYAN}$CLI_NAME MyGame${NC}"
     echo ""
     echo -e "  ${CYAN}# Create iOS project${NC}"
-    echo -e "  ${CYAN}$CLI_NAME create MyGame --platform ios --team-id YOUR_TEAM_ID${NC}"
+    echo -e "  ${CYAN}$CLI_NAME MyGame --platform ios --team-id YOUR_TEAM_ID${NC}"
     echo ""
     echo -e "  ${CYAN}# Create multi-platform project (macOS, iOS, visionOS)${NC}"
-    echo -e "  ${CYAN}$CLI_NAME create MyGame --platform multi --team-id YOUR_TEAM_ID${NC}"
+    echo -e "  ${CYAN}$CLI_NAME MyGame --platform multi --team-id YOUR_TEAM_ID${NC}"
     echo ""
 else
     echo -e "${YELLOW}⚠️  Warning: $CLI_NAME not found in PATH${NC}"
